+++ CHANGELOG.md
@@ -92,6 +92,8 @@
 72. Fixed Issue with MelonLoader Unity Debug Log not running when Plugins are loaded.
 73. Fixed Issue with VRChat_OnUiManagerInit not getting called.
 74. MelonLoader Unity Debug Log is now default even when no Plugins or Mods are loaded.
+75. Fixed Issue with No Attribute showing as Universal.
+76. Implemented Fallback Game Compatibility for Melons when no app.info is found.
 
 ---